requirements.txt
anyjson==0.3.3
boto==2.40.0
celery==3.1.23
Django==1.9.7
django-cors-headers==1.1.0
django-debug-toolbar==1.7
django-storages==1.5.0
django-user-sessions==1.3.0
djangorestframework==3.4.6
gunicorn==19.5.0
Markdown==2.6.6
msgpack-python==0.4.7
phonenumbers==7.4.0
raven==6.0.0
requests==2.11.1
factory_boy==2.8.1
django-anymail[sendgrid]==0.8
psycopg2==2.6.2
django-enumfields==0.9.0
asteval==0.9.7
rehive==0.3.6
twilio==5.4.0